requirements.txt
groq
langchain
langchain-community
langchain-huggingface
langchain-groq
langgraph
sentence-transformers
torch
chromadb
langchain-chroma
tavily-python
newspaper3k
beautifulsoup4
lxml
lxml-html-clean
requests
easyocr
pytesseract
Pillow
fastapi
uvicorn
python-multipart
streamlit
pydantic
pydantic-settings
python-dotenv
ragas
datasets
langsmith
pandas
numpy
tqdm
httpx